portfolio_rag/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
langchain-chroma>=1.0.0
langchain-huggingface>=1.2.0
langchain-community>=0.4.1
langchain-text-splitters>=1.0.0
langchain-core>=1.2.0
groq>=0.4.1
chromadb>=1.3.6
sentence-transformers>=2.2.2
pydantic>=2.5.0
numpy>=2.0.0

